Flatten the route ceremony
Eight files stood between main.tsx and the first content, serving a
router whose semantic load is one path param and query strings. The app
never had file-based routing — no codegen reads the routes/ layout; the
tree was always hand-assembled — so the scattered definition files were
convention worn as costume.

- The shell becomes Page.tsx exporting Page: it was an FC named Route,
  forcing route.tsx (exporting the real route object, also Route) to
  import it under a third name. The inner hub is PageBody; its css
  pairs as Page.module.css
- All four route definitions + RootLayout + the assembled tree collapse
  into one src-root routes.tsx: the redirector, the $id guard with its
  watcher lifecycle, the workspace route with its total-decoding
  validateSearch. routeTree.ts and the scattered files are deleted; the
  routes/ directory keeps mirroring the URL but holds only components
- IndexPage moves to routes/IndexPage.tsx as one of those components —
  it grew a real body in the onboarding work, so it keeps its own file
  while its route definition joins the others
- router.ts still imports only the RouteTree type, so the import-leaf
  rule that protects HMR is intact
- A trailhead comment atop Page.tsx points exploration up (bootstrap,
  providers, chrome) and down (address spaces, pages, Details)

The ladder reads main -> App -> routes -> Page: four hops, each with
something to say. No behavior change.
