--- apps/lite/ui/src/routes/RootLayout.tsx
@@ -1,13 +1,0 @@
-import { HotkeysProvider } from "@tanstack/react-hotkeys";
-import { Outlet } from "@tanstack/react-router";
-import type { FC } from "react";
-import styles from "./RootLayout.module.css";
-
-export const RootLayout: FC = () => (
-	<HotkeysProvider>
-		<div className={styles.dragRegion} />
-		<main className={styles.content}>
-			<Outlet />
-		</main>
-	</HotkeysProvider>
-);
