--- apps/lite/ui/src/routes/__root.tsx
@@ -1,11 +1,0 @@
-import type { QueryClient } from "@tanstack/react-query";
-import { createRootRouteWithContext } from "@tanstack/react-router";
-import { RootLayout } from "./RootLayout.tsx";
-
-interface RouteContext {
-	queryClient: QueryClient;
-}
-
-export const Route = createRootRouteWithContext<RouteContext>()({
-	component: RootLayout,
-});
